+++ Changelog.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  * Mixed rust/python layout
  * Added PEP 517 support
  * Added a `pyo-pack sdist` command as workaround for [pypa/pip#6041](https://github.com/pypa/pip/issues/6041)
+ * Support settings all applicable fields from the python core metadata specification in Cargo.toml
 
 ## [0.6.1]